# room-service/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API room-service",
        "description": "Dokumentasi API room-service menggunakan Swagger",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/rooms": {
            "get": {
                "tags": [
                    "Room"
                ],
                "summary": "Get list of all rooms",
                "operationId": "bea681612754126f01807ff95647eb34",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Room"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Room"
                ],
                "summary": "Create a new room",
                "operationId": "cd6c1029f22e2425a1f99cf83e27383a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoomCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Room created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rooms/{id}": {
            "get": {
                "tags": [
                    "Room"
                ],
                "summary": "Get room by ID",
                "operationId": "91a59d880bb0297e8321dd75af8f506b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Room ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room data retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Room not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Room"
                ],
                "summary": "Update room data",
                "operationId": "c489811fdd25dfe273cf9a5bc3b5d386",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Room ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoomUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Room updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Room not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Room"
                ],
                "summary": "Delete a room",
                "operationId": "61024891be80949c0aaa6e4536381225",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Room ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Room deleted"
                    },
                    "404": {
                        "description": "Room not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Room": {
                "required": [
                    "id",
                    "name",
                    "capacity"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Meeting Room A"
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 10
                    },
                    "facilities": {
                        "type": "string",
                        "example": "AC, Projector",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "RoomCreate": {
                "required": [
                    "name",
                    "capacity"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "New Meeting Room"
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 15
                    },
                    "facilities": {
                        "type": "string",
                        "example": "AC, Whiteboard",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "RoomUpdate": {
                "required": [
                    "capacity"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Updated Room Name"
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 20
                    },
                    "facilities": {
                        "type": "string",
                        "example": "AC",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "tags": [
        {
            "name": "Room",
            "description": "Room"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ]
}